+++ src/repositories/index.ts
@@ -4,3 +4,4 @@ export { missedRepository } from './missed';
 export { announcementsRepository } from './announcements';
 export { usersRepository } from './users';
 export { bandsRepository } from './bands';
+export { duckRepository } from './duck';